sexto commit

+++ backend/src/main/java/com/backend/crm/repositorio/AdmRepositorio.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdmRepositorio extends JpaRepository<Adm,Long>{
+public interface AdmRepositorio extends JpaRepository<Adm,Integer>{
     
     @Query("select a from Adm a where a.nomeUsuario = :nomeUsuario and a.senha = :senha")
     public Adm buscarLogin(String nomeUsuario, String senha);
