+++ backend/src/main/java/com/backend/crm/repositorio/LivrosRepositorio.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LivrosRepositorio extends JpaRepository<LivrosAr, Long>{
+public interface LivrosRepositorio extends JpaRepository<LivrosAr, Integer>{
     
     @Query("select a from LivrosAr a where a.nomeLivro = :nomeLivro")
     public LivrosAr findByNomeLivro(String nomeLivro);
